==> notebooks/06.ipynb <==
import apache_beam as beam
from apache_beam.options.pipeline_options import PipelineOptions
import csv

dataset1 = read_csv_file('../input_data/dataset1.csv')
dataset2 = read_csv_file('../input_data/dataset2.csv')

def read_csv_file(file_path):
    with open(file_path, newline='') as csvfile:
        reader = csv.reader(csvfile)
        next(reader)  # Skip header row
        return [tuple(row) for row in reader]

def merge_datasets(element):
    invoice_id, legal_entity, counter_party, rating, status, value, tier = element
    return legal_entity, counter_party, tier, rating, status, value

def filter_arap(element):
    legal_entity, counter_party, tier, rating, status, value = element
    return status == 'ARAP'

def filter_accr(element):
    legal_entity, counter_party, tier, rating, status, value = element
    return status == 'ACCR'

def sum_arap_values(elements):
    legal_entity, counter_party, tier, ratings, arap_values, _ = zip(*elements)
    return legal_entity[0], counter_party[0], tier[0], max(ratings), sum(arap_values), 0

def sum_accr_values(elements):
    legal_entity, counter_party, tier, ratings, _, accr_values = zip(*elements)
    return legal_entity[0], counter_party[0], tier[0], max(ratings), 0, sum(accr_values)

def write_to_csv(elements):
    with open('../output_data/output.csv', 'w', newline='') as csvfile:
        writer = csv.writer(csvfile)
        writer.writerow(['legal_entity', 'counter_party', 'tier', 'max_rating', 'arap_sum', 'accr_sum'])
        for element in elements:
            writer.writerow(element)

with beam.Pipeline(options=PipelineOptions()) as pipeline:
    input1 = pipeline | 'Create dataset1' >> beam.Create(dataset1)
    input2 = pipeline | 'Create dataset2' >> beam.Create(dataset2)

    merged = (
        {'dataset1': input1, 'dataset2': input2}
        | 'CoGroupByKey' >> beam.CoGroupByKey()
        | 'FlatMap Merged Datasets' >> beam.FlatMap(lambda elem: [(elem[0], *value) for value in elem[1]['dataset1']])
        | 'Merge Datasets' >> beam.Map(merge_datasets)
    )

    arap_sums = (
        merged
        | 'Filter ARAP' >> beam.Filter(filter_arap)
        | 'Group ARAP' >> beam.GroupBy(lambda elem: (elem[0], elem[1], elem[2]))
        | 'Sum ARAP Values' >> beam.Map(sum_arap_values)
    )

    accr_sums = (
        merged
        | 'Filter ACCR' >> beam.Filter(filter_accr)
        | 'Group ACCR' >> beam.GroupBy(lambda elem: (elem[0], elem[1], elem[2]))
        | 'Sum ACCR Values' >> beam.Map(sum_accr_values)
    )

    result = (
        (arap_sums, accr_sums)
        | 'Flatten Results' >> beam.Flatten()
        | 'Group Results' >> beam.GroupBy(lambda elem: (elem[0], elem[1], elem[2]))
        | 'Combine Results' >> beam.Map(lambda elem: (elem[0], elem[1], elem[2], max([x[3] for x in elem[1]]), sum([x[4] for x in elem[1]]), sum([x[5] for x in elem[1]])))
    )

    output = result | 'Collect Output' >> beam.CombineGlobally(beam.combiners.ToListCombineFn())
    output | 'Write to CSV' >> beam.Map(write_to_csv)
