==> notebooks/07.ipynb <==
import apache_beam as beam

with beam.Pipeline() as pipeline:
    lines = pipeline | 'Read CSV file' >> beam.io.textio.ReadFromText('../input_data/dataset1.csv')
    _ = lines | 'Print CSV data' >> beam.Map(print)
